Revert library changes

+++ libs/uart/api.h
@@ -3,11 +3,8 @@
 
 /*
  * Initializes UART peripheral with the specified baudrate
- *
- * If use_interrupt is true, UART receive interrupts will be enabled and the ISR
- * will be available using:
  */
-void uart_init(uint16_t baudrate);
+void uart_init(uint32_t baudrate);
 
 /*
  * Returns a single character from the UART peripheral
